+++ packages/build-tools/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Gradle plugin resolves artifacts with latest revision of version instead of exact version (#421)
+
 ## [0.2.1] - 2026-07-16
 
 ### Fixed